requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
configparser==5.3.0
confluent-kafka==2.1.1
idna==3.4
requests==2.31.0
urllib3==2.0.3
